PIPE-0.5 fix: bronze_playbyplay.season BIGINT not INT
Spark infers Python int → LongType in createDataFrame, but the DDL
declared season as INT (IntegerType). The first append failed with
DELTA_FAILED_TO_MERGE_FIELDS / IntegerType vs LongType. BIGINT in
both DDL and inferred Spark schema, so they reconcile. Also matches
Postgres bigint cleanly for the future migration.

+++ jobs/ingest/fetch_playbyplay.py
@@ -61,7 +61,7 @@ def _retry_sql(sql, attempts=6, initial_backoff_s=5.0):
 _retry_sql(f"""
     CREATE TABLE IF NOT EXISTS {fq}.bronze_playbyplay (
         game_pk BIGINT,
-        season INT,
+        season BIGINT,
         fetched_at_utc TIMESTAMP,
         payload_gz BINARY
     ) USING DELTA
